Prepare HashStore for factory design implementation by refactoring 'FileHashStore' to implement 'hashstore_interface' and removing the 'hashstore_interface' from HashStore (and updating pytests)
